requirements.txt
# ----------------------------
# Core scientific stack
# ----------------------------
numpy>=1.23,<3.0
pandas>=1.5,<3.0
scipy>=1.10,<2.0

# ----------------------------
# ML (Edge)
# ----------------------------
scikit-learn>=1.2,<2.0
joblib>=1.2,<2.0

# Imbalance handling (SMOTE/ENN/LOF uses sklearn too)
imbalanced-learn>=0.11,<1.0

# ----------------------------
# Deep Learning (Cloud)
# ----------------------------
tensorflow-cpu>=2.12,<2.18

# ----------------------------
# Plotting
# ----------------------------
matplotlib>=3.7,<4.0

# ----------------------------
# Optional quality-of-life (recommended)
# ----------------------------
tqdm>=4.66,<5.0
